Added a pixel padding to emojis in order to avoid bleeding into neighbors.
This guarantees that each image has 1px of transparent border inside itself.

This just fixes the minor inaccuracies when the game renders the images and floating numbers.

It's minor and not noticeable but it looks better now with this change.

The conversion seem to work fine, all that remains is to improve the unicode handling in the library as it can't properly match the full unicode string and will match each individual set of bytes and ignore modifiers like the more complex emojis might use.